+++ addOns/openapi/src/test/java/org/zaproxy/zap/extension/openapi/ExtensionOpenApiTest.java
@@ -58,11 +58,11 @@ import org.parosproxy.paros.db.TableHistory;
 import org.parosproxy.paros.extension.ExtensionLoader;
 import org.parosproxy.paros.model.HistoryReference;
 import org.parosproxy.paros.model.Model;
+import org.zaproxy.addon.commonlib.DefaultValueProvider;
 import org.zaproxy.addon.commonlib.ExtensionCommonlib;
 import org.zaproxy.zap.extension.ascan.VariantFactory;
 import org.zaproxy.zap.extension.openapi.OpenApiExceptions.InvalidDefinitionException;
 import org.zaproxy.zap.model.Context;
-import org.zaproxy.zap.model.DefaultValueGenerator;
 import org.zaproxy.zap.testutils.NanoServerHandler;
 import org.zaproxy.zap.utils.I18N;
 import org.zaproxy.zap.utils.ZapXmlConfiguration;
@@ -91,7 +91,7 @@ class ExtensionOpenApiTest extends AbstractServerTest {
         ExtensionCommonlib extCommonlib =
                 mock(ExtensionCommonlib.class, withSettings().strictness(Strictness.LENIENT));
         given(extensionLoader.getExtension(ExtensionCommonlib.class)).willReturn(extCommonlib);
-        given(extCommonlib.getValueGenerator()).willReturn(new DefaultValueGenerator());
+        given(extCommonlib.getValueProvider()).willReturn(new DefaultValueProvider());
 
         tableHistory = mock(TableHistory.class);
         HistoryReference.setTableHistory(tableHistory);