+++ addOns/formhandler/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- Depend on Common Library add-on, to provide the default/custom values to the other add-ons (Issue 8016).
 
 ## [6.6.0] - 2024-05-07
 ### Changed